+++ api/lib/application/organizations/index.js
@@ -174,47 +174,6 @@ const register = async function (server) {
     },
     {
       method: 'GET',
-      path: '/api/admin/organizations/{id}/memberships',
-      config: {
-        pre: [
-          {
-            method: (request, h) =>
-              securityPreHandlers.hasAtLeastOneAccessOf([
-                securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkAdminMemberHasRoleCertif,
-                securityPreHandlers.checkAdminMemberHasRoleSupport,
-                securityPreHandlers.checkAdminMemberHasRoleMetier,
-              ])(request, h),
-            assign: 'belongsToOrganization',
-          },
-        ],
-        validate: {
-          params: Joi.object({
-            id: identifiersType.organizationId,
-          }),
-          query: Joi.object({
-            filter: Joi.object({
-              firstName: Joi.string().empty('').allow(null).optional(),
-              lastName: Joi.string().empty('').allow(null).optional(),
-              email: Joi.string().empty('').allow(null).optional(),
-              organizationRole: Joi.string().empty('').allow(null).optional(),
-            }).default({}),
-            page: Joi.object({
-              number: Joi.number().integer().empty('').allow(null).optional(),
-              size: Joi.number().integer().empty('').allow(null).optional(),
-            }).default({}),
-          }),
-        },
-        handler: organizationController.findPaginatedFilteredMembershipsForAdmin,
-        tags: ['api', 'organizations'],
-        notes: [
-          'Cette route est restreinte aux utilisateurs de Pix Admin',
-          'Elle retourne les rôles des membres rattachés à l’organisation de manière paginée.',
-        ],
-      },
-    },
-    {
-      method: 'GET',
       path: '/api/admin/organizations/{organizationId}/children',
       config: {
         pre: [
