+++ api/tests/team/acceptance/application/membership/membership.admin.route.test.js
@@ -8,7 +8,7 @@ import {
   generateValidRequestAuthorizationHeader,
 } from '../../../../test-helper.js';
 
-describe('Acceptance | Controller | membership-controller', function () {
+describe('Acceptance | Team | Admin | Route | membership', function () {
   let server;
 
   beforeEach(async function () {
@@ -232,6 +232,67 @@ describe('Acceptance | Controller | membership-controller', function () {
         expect(response.statusCode).to.equal(400);
         const firstError = response.result.errors[0];
         expect(firstError.detail).to.equal('"id" must be a number');
+      });
+    });
+  });
+
+  describe('GET /api/admin/organizations/{id}/memberships', function () {
+    it('returns the matching membership as JSON API', async function () {
+      // given
+      const userSuperAdmin = databaseBuilder.factory.buildUser.withRole();
+      const organization = databaseBuilder.factory.buildOrganization();
+      const user = databaseBuilder.factory.buildUser();
+      const membershipId = databaseBuilder.factory.buildMembership({
+        userId: user.id,
+        organizationId: organization.id,
+      }).id;
+
+      await databaseBuilder.commit();
+
+      // when
+      const response = await server.inject({
+        method: 'GET',
+        url: `/api/admin/organizations/${organization.id}/memberships?filter[email]=&filter[firstName]=&filter[lastName]=&filter[organizationRole]=`,
+        headers: { authorization: generateValidRequestAuthorizationHeader(userSuperAdmin.id) },
+      });
+
+      // then
+      expect(response.statusCode).to.equal(200);
+      expect(response.result).to.deep.equal({
+        data: [
+          {
+            attributes: {
+              'organization-role': 'MEMBER',
+            },
+            id: membershipId.toString(),
+            relationships: {
+              user: {
+                data: {
+                  id: user.id.toString(),
+                  type: 'users',
+                },
+              },
+            },
+            type: 'organization-memberships',
+          },
+        ],
+        included: [
+          {
+            attributes: {
+              email: user.email,
+              'first-name': user.firstName,
+              'last-name': user.lastName,
+            },
+            id: user.id.toString(),
+            type: 'users',
+          },
+        ],
+        meta: {
+          page: 1,
+          pageCount: 1,
+          pageSize: 10,
+          rowCount: 1,
+        },
       });
     });
   });
